didn't realize shader header and source where flipped
